pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "ragamp"
version = "0.1.0"
authors = [
    {name = "Ozan Gokdemir"},
    {email = "ogokdemir@uchicago.edu"},
]
description = "Retrieval Augmented Generation for Antimicrobial Peptides."
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: Implementation :: CPython",
]

dependencies = [
    "requests >= 2.31.0",
    "transformers >= 4.37.1",
    "tokenizers >=  0.15.1",
    "llama-index >= 0.9.36",
    "langchain >= 0.1.3",
    "llama-index-llms-huggingface",
    "llama-index-readers-file",
    "bitsandbytes >= 0.42.0",
    "sentence_transformers >= 2.2.2",
    "llama-index-embeddings-langchain",
    "llama-index-embeddings-huggingface",
    "llama-index-vector-stores-faiss",
    "faiss-gpu >= 1.7.0",

]

[project.urls]
homepage = "https://github.com/ogkdmr/ragamp"
documentation = "https://ogkdmr.github.io/ragamp/"
repository = "https://github.com/ogkdmr/ragamp"

[project.optional-dependencies]
dev = [
    "covdefaults>=2.2",
    "coverage",
    "mypy",
    "pre-commit",
    "pytest",
    "pytest-cov",
    "ruff>=0.2.0",
    "tox",
    "virtualenv",
]
docs = [
    "black",
    "mkdocs-gen-files",
    "mkdocs-literate-nav",
    "mkdocs-material==9.4.7",
    "mkdocs-section-index",
    "mkdocstrings==0.23.0",
    "mkdocstrings-python==1.8.0",
    "mike",
]

[tool.codespell]
skip = """
.git,
.github,
__pycache__,
build,
dist,
.*egg-info
"""

[tool.coverage.run]
plugins = ["covdefaults"]
omit = ["examples"]

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "testing.*"
allow_incomplete_defs = true
allow_untyped_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
allow_incomplete_defs = true
allow_untyped_defs = true

[tool.ruff]
line-length = 79
target-version = "py38"

[tool.ruff.format]
indent-style = "space"
quote-style = "single"

[tool.ruff.lint]
# See all rules here: https://beta.ruff.rs/docs/rules
select = [
    # pyflakes
    "F",
    # pycodestyle
    "E",
    # mccabe
    "C90",
    # isort
    "I",
    # pep8-naming
    "N",
    # pydocstyle
    "D",
    # pyupgrade
    "UP",
    # flake8-2020
    "YTT",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-commas
    "COM",
    # flake8-comprehensions
    "C4",
    # flake8-implicit-str-concat
    "ISC",
    # flake8-pytest-style
    "PT",
    # flake8-quotes
    "Q",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # PyLint
    "PL",
    # ruff-specific
    "RUF",
]
extend-ignore = ["COM812", "ISC001"]

[tool.ruff.lint.flake8-pytest-style]
parametrize-values-type = "tuple"

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"
multiline-quotes = "double"

[tool.ruff.lint.isort]
force-single-line = true
known-first-party = ["ragamp", "test", "testing"]
order-by-type = false
required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.per-file-ignores]
"*/__init__.py" = ["F401"]
"*/*_test.py" = ["D10"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.setuptools.packages.find]
exclude = ["tests*", "testing*"]
namespaces = false
